+++ src/test/browser.js
@@ -30,10 +30,10 @@ export default async (argv, execaOptions) => {
   const files = argv.files.length > 0
     ? argv.files
     : [
-        'test/**/*.spec.{js,cjs,mjs}',
-        'test/browser.{js,cjs,mjs}',
-        'dist/test/**/*.spec.{js,cjs,mjs}',
-        'dist/test/browser.{js,cjs,mjs}'
+        'test/**/*.spec.*js',
+        'test/browser.*js',
+        'dist/test/**/*.spec.*js',
+        'dist/test/browser.*js'
       ]
 
   // before hook