+++ test/browser.spec.js
@@ -12,10 +12,12 @@ describe('browser', () => {
 
   it('should load tar.gz fixures', () => {
     const myFixture = loadFixture('test/fixtures/file.tar.gz')
+    // spell-checker: disable-next-line
     expect(uint8ArrayToString(myFixture, 'base64')).to.equal('H4sICIlTHVIACw')
   })
 
   it('non existing fixtures', () => {
+    // spell-checker: disable-next-line
     expect(() => loadFixture('/test/fixtures/asdalkdjaskldjatest.txt'))
       .to.throw()
   })