+++ test/fixtures/projects/a-multiple-output-project/src/sw.ts
@@ -1,0 +1,3 @@
+export function main (): string {
+  return 'hello world from sw!'
+}